sensor_node/pyproject.toml
[project]
name = "sensor-node"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.4"
dependencies = [
    "zaichenko-engineer-telemetry-community-nipunn1313-mypy-grpc==3.6.0.1.20250622183011+2e812e491c9b",
    "zaichenko-engineer-telemetry-grpc-python==1.73.0.1.20250622183011+2e812e491c9b",
]

[[tool.uv.index]]
url = "https://buf.build/gen/python"

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "ruff>=0.12.0",
]
